pyproject.toml
[project]
name = "mawa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.47.0",
    "mistralai>=1.9.11",
    "python-dotenv>=1.2.1",
    "pypdf>=5.0.0",
    "supabase>=2.22.3",
    "typer>=0.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mawa"]
